+++ deltachat-jsonrpc/src/api/types/mod.rs
@@ -6,6 +6,7 @@ pub mod location;
 pub mod message;
 pub mod provider_info;
 pub mod qr;
+pub mod reactions;
 pub mod webxdc;
 
 pub fn color_int_to_hex_string(color: u32) -> String {